% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_acs-tables.R
\docType{data}
\name{acs_NY}
\alias{acs_NY}
\title{Sample cumulative CCES data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4320 rows and 6 columns.
}
\source{
Kyle Walker and Matt Herman (2021). tidycensus: Load US Census Boundary and Attribute Data as
'tidyverse' and 'sf'-Ready Data Frames. R package version 0.11.4.
\url{https://CRAN.R-project.org/package=tidycensus}
}
\usage{
acs_NY
}
\description{
Table of counts by CD from the 2018 one-year ACS in New York. We load the
variables that roughly partition gender, age, and sex. That is, this is the
output of \code{get_acs_cces(acscodes_age_sex_race)}.
}
\examples{
 library(tibble)
 acs_NY

}
\keyword{datasets}
