% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acscodes-key.R
\docType{data}
\name{acscodes_age_sex_educ}
\alias{acscodes_age_sex_educ}
\title{ACS codes for a age-sex-education partition}
\format{
An object of class \code{character} of length 70.
}
\usage{
acscodes_age_sex_educ
}
\description{
ACS codes for a age-sex-education partition
}
\details{
Each element of the vector is a code for a ACS variable, which can be used
for example  in the \code{variables} argument of \code{tidycensus::get_acs()}.
There are 70 codes in \code{acscodes_age_sex_edu} because they specify cells
interacting age (5 bins), sex (2 bins), and education (7 bins). Some of
these will be collapsed for MRP to form a common denominator with the CCES; see
\link{educ_key}
}
\examples{
head(acscodes_age_sex_educ)
}
\seealso{
\link{acscodes_age_sex_rac}
}
\keyword{datasets}
