% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acscodes-key.R
\docType{data}
\name{acscodes_age_sex_race}
\alias{acscodes_age_sex_race}
\title{ACS codes for a age-sex-race partition}
\format{
An object of class \code{character} of length 160.
}
\usage{
acscodes_age_sex_race
}
\description{
ACS codes for a age-sex-race partition
}
\details{
Each element of the vector is a code for a ACS variable, which can be used,
for example  in the \code{variables} argument of \code{tidycensus::get_acs()}.
There are 160 codes in \code{acscodes_age_sex_rac} because they specify cells
interacting age (10 bins), sex (2 bins), and race/ethnicity (8 bins).  Some of
these will be collapsed for MRP to form a common denominator with the CCES; see
\link{race_key}
}
\examples{
head(acscodes_age_sex_race)
}
\seealso{
\link{acscodes_age_sex_educ}
}
\keyword{datasets}
