% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acscodes-key.R
\docType{data}
\name{acscodes_df}
\alias{acscodes_df}
\title{ACS code lookup table}
\format{
A data frame where each row represents a column.
\describe{
\item{variable}{the ACS code (2018)}
\item{gender}{A labelled variable for gender. 1 is Male, 2 is Female. Use
the \code{labelled} or \code{haven} package to see labels.}
\item{age}{A labelled variable specifying which age biin the variable specifies}
\item{race}{A labelled variable specifying which education bin the variable specifies}
\item{educ}{A labelled variable specifying which race binthe variable specifies}
}
}
\source{
Modifications around tidycensus::load_variables
}
\usage{
acscodes_df
}
\description{
A tidy dataframe where each row is a ACS code. This is useful internal
data to give meaning to variable codes e.g. \link{acscodes_age_sex_educ}
}
\details{
The 5-yr ACS at 2018 is used,
although codes should be fairly consistent across time. IF a demographic variable is NA,
that means the variable collapses over the levels of that variable. In other
wods, NA here can be thought of as meaning "all".
}
\examples{
 head(acscodes_df)

}
\keyword{datasets}
