% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acscodes-key.R
\docType{data}
\name{acscodes_partitions}
\alias{acscodes_partitions}
\alias{acscodes_age_sex_race}
\alias{acscodes_age_sex_educ}
\alias{acscodes_sex_educ_race}
\title{ACS codes for partitions}
\format{
\subsection{acscodes_age_sex_race}{

There are 160 codes in \code{acscodes_age_sex_race} because they specify cells
interacting age (10 bins), sex (2 bins), and race/ethnicity (8 bins).
}

\subsection{acscoes_age_sex_educ}{

There are 70 codes in \code{acscodes_age_sex_educ} because they specify cells
interacting age (5 bins), sex (2 bins), and education (7 bins).
}

\subsection{acscodes_sex_educ_race}{

There are 48 codes in \code{acscodes_sex_educ_race} because they specify cells
interacting sex (2 bins), education (3 bins), and race (8 bins). The entire partition
is not actually exhaustive; it appears to only limit to 25 years and above
and not incude postgraduate degrees. Cross-check with \link{acscodes_df} to verify.
}
}
\usage{
acscodes_age_sex_race

acscodes_age_sex_educ

acscodes_sex_educ_race
}
\description{
Each vector defines a partition, or at least an attempt at one. Each element of
the vector is a code for a ACS variable, which can be used,
for example in the \code{variables} argument of \code{tidycensus::get_acs()}. Some of
these will be collapsed for MRP to form a common denominator with the CCES; see
\link{keyvalue}
}
\examples{
head(acscodes_age_sex_race)
head(acscodes_age_sex_educ)
head(acscodes_sex_educ_race)
}
\seealso{
\link{acscodes_df} for a definition of all codes,
and \url{https://www.shirokuriwaki.com/ccesMRPprep/articles/acs.html} for a Vignette
on ACS value pairs
}
\keyword{datasets}
