% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_cces-common-samp.R
\docType{data}
\name{cc18_NY}
\alias{cc18_NY}
\title{Sample 2018 New York CCES}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3455 rows and 24 columns.
}
\source{
Kuriwaki, Shiro, 2018, "Cumulative CCES Common Content (2006-2019)",
\url{https://doi.org/10.7910/DVN/II2DB6}, Harvard Dataverse
}
\usage{
cc18_NY
}
\description{
Unlike \code{cc18_samp}, this dataset is not a sample and takes all the respondents
from New York in the 2018 CCES. It has 3455 rows with 24
columns.
}
\keyword{datasets}
