% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cces_std-for-acs.R
\name{ccc_bin_age}
\alias{ccc_bin_age}
\title{Discretize a vector of age integers into labelled variables}
\usage{
ccc_bin_age(agevec, agelbl = deframe(ccesMRPprep::age5_key))
}
\arguments{
\item{agevec}{a vector of integers}

\item{agelbl}{a value-key pair to be passed into recode,
with values as the things to be recoded and labels as the labels for each
value.}
}
\description{
Discretize a vector of age integers into labelled variables
}
\details{
The recoding is governed by \code{age5_key}. IT currently only accepts
5-way binning, following the ACS.
}
\examples{
  ccc_bin_age(c(15:100, NA))

}
