% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cces_std-for-acs.R
\name{ccc_std_demographics}
\alias{ccc_std_demographics}
\title{Recode CCES variables so that they merge to ACS variables}
\usage{
ccc_std_demographics(tbl, only_demog = FALSE)
}
\arguments{
\item{tbl}{A subset of the cumulative common content. Must include variables
\code{age}, \code{race}, \code{educ}, \code{gender}, \code{st}, \code{state},
and \code{cd}. See \link{ccc_samp} for an example.}

\item{only_demog}{Drop variables besides demographics? Defaults to FALSE}
}
\value{
The output is of the same dimensions as the input (unless \code{only_demog = TRUE})
but with the following exceptions:
\itemize{
\item \code{age_bin} is coded to match up with the ACS bins and the recoding occurs
in a separate function, \code{ccc_bin_age}.
\item \code{educ} is recoded (coarsened and relabelled) to match up with the ACS.
(the original version is left as \code{educ_cces_chr}). Recoding is governed by
the key-value pairs \link{educ_key}
\item the same goes for \code{race}. These recodings are governed by the
key-value pair \link{race_key}.
\item \code{cd} is standardized so that at large districs are given "01" and
single-digit districts are padded with 0s. e.g. \code{"WY-01"} and \code{"CA-02"}.
}
}
\description{
Recode CCES variables so that they merge to ACS variables
}
\examples{
\dontrun{
 library(dataverse)

 # Load Cumulative dataset (need your own Token)
 cumulative_rds <- get_cces_dv("cumulative")

 cumulative_std <- ccc_std_demographics(cumulative_rds)

 }

}
