% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs.R
\name{cd_from_acs}
\alias{cd_from_acs}
\title{Format CD format from ACS}
\usage{
cd_from_acs(vec)
}
\arguments{
\item{vec}{a vector of strings from the ACS congressional district naming}
}
\description{
Convert ACS' cd labels to the form "AK-01".
}
\examples{
 cd_from_acs("Congressional District 32 (115th Congress), California")


}
