% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_cd-info-long.R
\docType{data}
\name{cd_info_long}
\alias{cd_info_long}
\title{Congressional District level information - long version}
\format{
\code{cd_info_long} is a dataframe with 13050 rows,
covering the maps of 9 election years
(2008, 2010, 2012, 2014, 2016, 2018, 2020, 2022, 2024) for each of the 435 congressional districts.
\describe{
\item{\code{lines}}{Is the year corresponding to the geography (district line). For example, \code{lines = 2008} and \verb{cd = "AL-01}
indicates that the row is representing AL-01's geography as used in the 2008 election.}
\item{\code{cd}}{Is the CD corresponding to the year of the geography (district line). Note that districts can change drastically
by redistricting; a state's "first congressional district" from one \code{lines} can cover a different area
than the same first congressional district for another.}
\item{\code{elec}}{Is the year of the election for the presidential election data that follows}
\item{\code{party}, \code{candidate}}{Define the presidential candidate that corresponds to the \code{elec}
(which may not be the same as \code{lines}). For example, \verb{lines = 2012, cd = AL-01} combined with
\code{elec = 2008} represents the 2008 election results in the newly redistricted (2012) AL-01 geography}
\item{\code{pct}}{The two party voteshares of the candidate}
\item{\code{presvotes_total}}{The total number of votes for President in that CD}
\item{\code{presvotes_DR}}{The total number of Democrat + Republican votes for President in that CD}
}
}
\usage{
cd_info_long
}
\description{
\code{cd_info_long} provides a "long" version of the yearly \verb{cd_info_20**} datasets.
}
\examples{
 library(dplyr)

 # get only data for proximate years
 cd_info_long |> filter((elec == lines) | (elec + 2 == lines))

 # this subset returns exactly 2 * 435 districts per cycle:
 cd_info_long |> filter((elec == lines) | (elec + 2 == lines)) |> count(lines, party)

# this will show where the districts lines changed between 2022 and 2024
# (same election, same candidate, different map)
cd_info_long |>
 filter(lines \%in\% c(2022, 2024), elec == 2020, candidate == "biden") |>
 arrange(cd, lines)

}
\seealso{
\link{cd_info} for versions with one row per district, and documentation of sources
}
\keyword{datasets}
