% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_cd-info.R
\docType{data}
\name{cd_presvote}
\alias{cd_presvote}
\title{Presidential Voteshare by Daily Kos}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3045 rows and 4 columns.
}
\source{
Daily Kos, "2008, 2012, & 2016 results for districts used in 2018."
. \url{https://bit.ly/3DRhPcj}
}
\usage{
cd_presvote
}
\description{
Similar format to \verb{cd_info_20**} but stacked from 2006 to 2018. Focuses only on voteshare.
}
\keyword{datasets}
