% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_cps-onewaytabs.R
\docType{data}
\name{cps_onewaytabs}
\alias{cps_onewaytabs}
\title{CPS variable labels and counts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3258 rows and 5 columns.
}
\usage{
cps_onewaytabs
}
\description{
One summary of the CPS, by reporting all year-by-year raw counts of the microdata
for main demographic varibales.
}
\examples{
head(cps_onewaytabs)
}
\seealso{
The \href{https://www.shirokuriwaki.com/ccesMRPprep/articles/cps.html}{Vignette on CPS}.
}
\keyword{datasets}
