% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cces_format.R
\name{join_slim}
\alias{join_slim}
\title{Join and slim person-level dataset with question-level dataset}
\usage{
join_slim(ccq_df, ccc_df, cd_df, formula, subset_dist = NA)
}
\arguments{
\item{subset_dist}{a character for the geography of cd to subset}
}
\description{
Join and slim person-level dataset with question-level dataset
}
