% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_keyvalues.R
\docType{data}
\name{keyvalue}
\alias{keyvalue}
\alias{race_key}
\alias{gender_key}
\alias{educ_key}
\alias{educ_key2}
\alias{age5_key}
\alias{age10_key}
\title{CCES-ACS variable key value pairs for recoding}
\format{
All keys are tibbles with one row per recoding value.

\subsection{\code{race_key}}{

\describe{
\item{race}{An labelled integer of class haven::labelled.
Most compact form of both sources and the values both will get recoded
to in MRP.}
\item{race_cces}{Labelled versions of the CCES race codings. These are of the
same class as the CCES cumualtive file.}
\item{race_cces_chr}{Labels for the first column, in characters}
\item{race_acs}{Corresponding character in the ACS data via the tidycensus package}
\item{race_5}{A numeric value underlying the \code{race} label.}
}
}

\subsection{\code{gender_key}:}{

\describe{
\item{gender}{An labelled integer of class haven::labelled. Target variable}
\item{gender_chr}{Character to recode from. CCES and ACS use the same values.}
}
}

\subsection{\code{educ_key}}{

For mapping ACS data values for education e.g. in \link{get_acs}:
\describe{
\item{educ_cces_chr}{Character to recode from, in CCES}
\item{educ_chr}{Character to recode from, in ACS. See \link{}}
\item{educ}{An labelled integer of class haven::labelled. Target variable}
}
}

\subsection{\code{educ_key2}}{

Education with slightly different ACS value encoding, used e.g. in building
\link{acscodes_age_sex_educ}:
\describe{
\item{educ_cces_chr}{Character to recode from, in CCES}
\item{educ_chr}{Character to recode from, in ACS. See \link{}}
\item{educ}{An labelled integer of class haven::labelled. Target variable}
}
}

\subsection{\code{age5_key}}{

Age bins, 5-ways, used in \link{acscodes_age_sex_educ}. Use \link{ccc_bin_age}
to recode CCES variabe
\describe{
\item{age}{An labelled integer of class haven::labelled. Target variable.}
\item{age_chr}{Character to recode from, in ACS}
}
}

\code{age10_key}: Age bins, 10-ways, used in \link{acscodes_age_sex_race}:
\describe{
\item{age}{An labelled integer of class haven::labelled. Target variable.}
\item{age_chr}{Character to recode from, in ACS}
}
}
\usage{
race_key

gender_key

educ_key

educ_key2

age5_key

age10_key
}
\description{
CCES-ACS variable key value pairs for recoding
}
\details{
These value-value tables are useful for recoding the values of
from one dataset (CCES) so that they can be merged immediately with another
(ACS). These get used internally in \link{ccc_std_demographics}, but they are
available as built in datasets
}
\examples{
 library(ccesMRPprep)
 race_key
 educ_key
 educ_key2
 gender_key
 age5_key
 age10_key
}
\keyword{datasets}
