% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs-poststrat.R
\name{std_acs_cdformat}
\alias{std_acs_cdformat}
\title{Format CD format from ACS}
\usage{
std_acs_cdformat(vec)
}
\arguments{
\item{vec}{a vector of strings from the ACS congressional district naming}
}
\description{
Convert ACS' cd labels to the form "AK-01".
}
\examples{
 std_acs_cdformat("Congressional District 32 (115th Congress), California")


}
