% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_bmlogit.R
\name{synth_bmlogit}
\alias{synth_bmlogit}
\title{Synthetic joint estimation with balancing}
\usage{
synth_bmlogit(
  formula,
  microdata,
  poptable,
  fix_to,
  area_var,
  count_var = "count",
  tol = 0.01
)
}
\arguments{
\item{fix_to}{A dataset with only marginal counts or proportions of the outcome
in question, by each area. Proportions will be corrected so that the margins
of the synthetic joint will match these, with a simple ratio.}

\item{tol}{Tolerance for balance}
}
\description{
Synthetic joint estimation with balancing
}
\examples{

educ_target <- count(acs_educ_NY, cd, educ, wt = count, name = "count")
pop_syn <- synth_onestep(educ ~ race + age + female,
                           microdata = cc18_NY,
                           fix_to = educ_target,
                           poptable = acs_race_NY,
                           area_var = "cd")

}
\seealso{
\code{synth_mlogit()}
}
\keyword{internal}
