% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_additional-X.R
\name{synth_mlogit}
\alias{synth_mlogit}
\title{Impute joint distribution of categorical variables}
\source{
Soichiro Yamauchi (2021). emlogit: Implementing the ECM algorithm for multinomial logit
model. R package version 0.1.1. \url{https://github.com/soichiroy/emlogit}

Yair Ghitza and Mark Steitz (2020). DEEP-MAPS Model of the Labor Force.
Working Paper. \url{https://github.com/Catalist-LLC/unemployment}
}
\usage{
synth_mlogit(formula, microdata, poptable, area_var, count_var = "count")
}
\arguments{
\item{formula}{A representation of the aggregate imputation or "outcome" model,
of the form \verb{X_\{K\} ~ X_1 + ... X_\{K - 1\}}}

\item{microdata}{The survey table that the multinomial model will be built off.
Must contain all variables in the LHS and RHS of \code{formula}.}

\item{poptable}{The population table, collapsed in terms of counts. Must contain
all variables in the RHS of \code{formula}, as well as the variables specified in
\code{area_var} and \code{count_var} below.}

\item{area_var}{A character vector of the area of interest.}

\item{count_var}{A character variable that specifies which variable in \code{poptable}
indicates the count}
}
\description{
Imputes the counts of a joint distribution of count variables for small areas
based on microdata. In this setup, the population distribution table (\code{poptable})
has the joint distribution of \verb{(A, X_\{1\}, ..., X_\{K - 1\})} categorical variables where
\code{A} denotes a categorical small area, \code{X}s denote categorical covariates, and
the missing covariate is \verb{X_\{K\}}.
}
\details{
Now, the survey data (\code{microdata}) has a sample joint distribution of
\verb{(X_\{1\}, .., X_\{K - 1\}, X_\{K\})} categorical variables but the sample size is too small
for small areas. Therefore, the function models a multinomial outcome model roughly of the
form \verb{X_\{K\} ~ X_\{1\} + ... X_\{K\}} and predicts onto \code{poptable} to estimate the
joint distribution of \verb{(A, X_\{1\}, .., X_\{K\})}
}
\examples{
 library(dplyr)

 # Impute the joint distribution of party ID with race, sex, and age, using
 # survey data in NY.

 imputed_acs <- synth_mlogit(pid3 ~ race + age + female,
                             microdata = cc18_NY,
                             poptable = acs_NY,
                             area_var = "cd")

 # original (27 districts x 2 sex x 5 age x 6 race categories)
 count(acs_NY, cd, female, age, race, wt = count)

 # new, modeled (original x 5 party categories)
 imputed_acs

 # See the data elec_NY to see if these numbers look reasonable.


\dontrun{
  # another example -- imputing education
  library(ccesMRPrun)
  synth_mlogit(educ ~ age + female,
              microdata = cces_GA,
              poptable = acs_GA,
              area_var = "cd")
}

}
