% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district-coding.R
\name{to_cd}
\alias{to_cd}
\title{Concatenate to Congressional District code}
\usage{
to_cd(state, num)
}
\arguments{
\item{state}{A vector of state names, preferably abbreviations}

\item{num}{A vector of district codes}
}
\description{
Concatenate to Congressional District code
}
\examples{
 to_cd(c("AL", "AK"), c("5", "AL"))

}
