% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_acscodes.R
\docType{data}
\name{acscodes_df}
\alias{acscodes_df}
\title{ACS code lookup table}
\format{
A data frame with 26,271 rows where
each row represents a variable in the ACS for which there is a count (e.g.
18-24 year olds who identify as Hispanic).

\describe{
\item{variable}{the ACS code (2018)}
\item{gender}{A labelled variable for gender. 1 is Male, 2 is Female. Use
the \code{labelled} or \code{haven} package to see labels.}
\item{age}{A labelled variable specifying which age bin the variable specifies}
\item{race}{A labelled variable specifying which education bin the variable specifies}
\item{educ}{A labelled variable specifying which race bin the variable specifies}
}
}
\source{
Modifications around \code{tidycensus::load_variables}
}
\usage{
acscodes_df
}
\description{
A tidy dataframe where each row is a ACS code. This is useful internal
data to give meaning to variable codes e.g. \link{acscodes_age_sex_educ}
}
\details{
The 5-yr ACS at 2018 is used,
although codes should be fairly consistent across time. IF a demographic variable is \code{NA},
that means the variable collapses over the levels of that variable. In other
words, \code{NA} here can be thought of as meaning "all".
}
\examples{
 head(acscodes_df)

}
\keyword{datasets}
