% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cces_collapse.R
\name{build_counts}
\alias{build_counts}
\title{Collapse CCES data to be analyzed in binomial model}
\usage{
build_counts(data, model_ff, multiple_qIDs = FALSE)
}
\arguments{
\item{data}{A cleaned CCES dataset, e.g. from \link{ccc_std_demographics} which is
then combined with outcome and contextual data in \link{cces_join_slim}. Currently, it
assumes the outcome is named \code{response}. This variable need not be numeric or
binarized, but it will be binarized with \link{yesno_to_binary}.}

\item{model_ff}{the model formula used to fit the multilevel regression model.
Currently only expects an binomial, of the brms form \code{y|trials(n) ~ x1 + x2 + (1|x3)}.
Only the RHS will be used but the LHS is necessary.}

\item{multiple_qIDs}{Does the data contain \emph{multiple} outcomes in long form and
therefore require the counts to be built for each outcome? Defaults to \code{FALSE}.}
}
\description{
Collapse CCES data to be analyzed in binomial model
}
\examples{

library(dplyr)

ccc_samp_std <- ccc_samp \%>\%
  mutate(response = sample(c("For", "Against"), size = n(), replace = TRUE)) \%>\%
  ccc_std_demographics()

ccc_samp_out <- build_counts(ccc_samp_std,
                             "yes | trials(n_response) ~ age + gender + educ")
}
