% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_for-acs.R
\name{ccc_std_demographics}
\alias{ccc_std_demographics}
\title{Recode CCES variable values so that they merge to ACS variables}
\usage{
ccc_std_demographics(tbl, only_demog = FALSE)
}
\arguments{
\item{tbl}{A subset of the cumulative common content. Must include variables
\code{age}, \code{race}, \code{educ}, and \code{gender}.}

\item{only_demog}{Drop variables besides demographics? Defaults to FALSE}
}
\description{
Recode CCES variable values so that they merge to ACS variables
}
\examples{
 library(dataverse)

 # Load Cumulative dataset
 cumulative_rds <- get_file("cumulative_2006_2018.Rds", "doi:10.7910/DVN/II2DB6")
 tmp <- tempfile(fileext = ".Rds")
 writeBin(cumulative_rds, tmp)
 cumulative_rds <- readr::read_rds(tmp)

 cumulative_std <- ccc_std_demographics(cumulative_rds)

}
