% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataverse.R
\name{get_cces_dv}
\alias{get_cces_dv}
\title{Download a specific CCES dataset from dataverse}
\usage{
get_cces_dv(
  name = "cumulative",
  year_subset = 2006:2019,
  dataverse_paths = cces_dv_ids
)
}
\arguments{
\item{name}{The name of the dataset as defined in \code{data(cces_dv_ids)}.}

\item{year_subset}{The year (or years, a vector) to subset too. If \code{name} is a year specific dataset, this
argument is redundant, but if \code{name == "cumulative"}, then the output
will be the cumulative dataset subsetted to that year. This is useful when
using the cumulative dataset for its harmonized variables.}

\item{dataverse_paths}{A dataframe where one row represents metadata for one
CCES dataset. Built-in data \code{cces_dv_ids} is used as a default and should
not be changed.}
}
\description{
You must download the development version of https://github.com/IQSS/dataverse-client-r
and get your own dataverse API Key. The current dataverse package downloads
the raw data, so this function writes the raw binary into a tempfile and
loads it into a tibble with the appropriate file data type.
}
\details{
To get a user key, you need to have your own dataverse account. Go to your account
online and look for "API Token". Generate a API Token. You can then save your
token by following \code{dataverse::get_user_key}, or directly saving
\code{Sys.setenv("DATAVERSE_KEY" = "YOUR KEY HERE")}
\code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.harvard.edu")}
}
\examples{

\dontrun{
 ccc <- get_cces_dv("cumulative", year = 2018)
 }

}
