% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataverse-pull.R
\name{get_cces_dv}
\alias{get_cces_dv}
\title{Download a specifc CCES dataset from dataverse}
\usage{
get_cces_dv(name = cces_dv_ids$cces_name, dataverse_paths = cces_dv_ids)
}
\arguments{
\item{name}{The name of the dataset as defined in \code{data(cces_dv_ids)}.}
}
\description{
You must download the development version of https://github.com/IQSS/dataverse-client-r
 and get your own dataverse API Key. The current dataverse package downloads
 the raw data, so this function writes the raw binary into a tempfile and
 loads it into a tibble with the appropriate file data type.
}
\details{
To get a user key, you need to have your own dataverse account. Go to your account
 online and look for "API Token". Generate a API Token. You can then save your
 token by following \code{dataverse::get_user_key}, or directly saving
 \code{Sys.setenv("DATAVERSE_KEY" = "YOUR KEY HERE")}
 \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.harvard.edu")}
}
\examples{
 ccc <- get_cces_dv("cumulative")

}
