% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome.R
\name{get_cces_question}
\alias{get_cces_question}
\title{Obtain the dataset that contains the question}
\usage{
get_cces_question(.qcode, .year, .q_df = q_df)
}
\description{
Obtain the dataset that contains the question
}
