% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_value-recode-key.R
\docType{data}
\name{race_key}
\alias{race_key}
\alias{gender_key}
\alias{educ_key}
\alias{age5_key}
\alias{age10_key}
\title{CCES-ACS variable recode keys}
\format{
A tibble with one row per recoding value
 \describe{
 \item{race}{An labelled integer of class haven::labelled. Most compact form
 of all the data.}
 \item{race_chr, gender_chr, educ_chr}{Labels, in characters}
 \item{race_cces, edu_cces}{Corresponding labelled in the CCES cumulative common content}
 \item{race_acs}{Corresponding character in the ACS data via the tidycensus package}
 }

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 3 columns.
}
\usage{
race_key

gender_key

educ_key

age5_key

age10_key
}
\description{
These value-value tables are useful for recoding the values of
from one dataset (CCES) so that they can be merged immediately with another (ACS).
}
\keyword{datasets}
