% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_bmlogit.R
\name{synth_bmlogit}
\alias{synth_bmlogit}
\title{Synthetic joint estimation with balancing}
\usage{
synth_bmlogit(
  formula,
  microdata,
  poptable,
  fix_to,
  fix_by_area = any(area_var \%in\% colnames(fix_to)),
  area_var,
  count_var = "count",
  tol = 0.05
)
}
\arguments{
\item{fix_to}{A dataset with only marginal counts or proportions of the outcome
in question, by each area. Proportions will be corrected so that the margins
of the synthetic joint will match these, with a simple ratio.}

\item{fix_by_area}{logical, whether to fix to targets area by area. Defaults to
TRUE if \code{area_var} is a variable in \code{fix_to}. If \code{FALSE}, collapses the
input to a single target.}

\item{tol}{Tolerance for balance}
}
\description{
Synthetic joint estimation with balancing
}
\examples{

# can take a few minutes if fix_by_area = TRUE (the default)
educ_target <- count(acs_educ_NY, cd, educ, wt = count, name = "count")
pop_syn <- synth_bmlogit(educ ~ race + age + female,
                         microdata = cc18_NY,
                         fix_to = educ_target,
                         poptable = acs_race_NY,
                         area_var = "cd")
}
\seealso{
\code{synth_mlogit()}
}
