% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_additional-X.R
\name{synth_mlogit}
\alias{synth_mlogit}
\title{Model synthetic joint distribution of categorical variables}
\source{
Soichiro Yamauchi (2021). emlogit: Implementing the ECM algorithm for multinomial logit
model. R package version 0.1.1. \url{https://github.com/soichiroy/emlogit}

Yair Ghitza and Mark Steitz (2020). DEEP-MAPS Model of the Labor Force.
Working Paper. \url{https://github.com/Catalist-LLC/unemployment}
}
\usage{
synth_mlogit(formula, microdata, poptable, area_var, count_var = "count")
}
\arguments{
\item{formula}{A representation of the aggregate imputation or "outcome" model,
of the form \verb{X_\{K\} ~ X_1 + ... X_\{K - 1\}}}

\item{microdata}{The survey table that the multinomial model will be built off.
Must contain all variables in the LHS and RHS of \code{formula}.}

\item{poptable}{The population table, collapsed in terms of counts. Must contain
all variables in the RHS of \code{formula}, as well as the variables specified in
\code{area_var} and \code{count_var} below.}

\item{area_var}{A character vector of the area of interest.}

\item{count_var}{A character variable that specifies which variable in \code{poptable}
indicates the count}
}
\value{
A dataframe with a similar format as the \code{poptarget} table
but with rows expanded to serve as a joint distribution. In general,
if the variable of interest has \code{L} values, the final dataset will have
\code{L} times more rows than \code{poptarget}. The data will have additional variables:
\itemize{
\item \code{n_aggregate}: The sum of counts known in the aggregate. i.e., the number of
trials the multinomial will consider. This is the sum of the original
\code{count} variables.
\item The outcome variable of interest. For example if the LHS of the formula
was \code{party_id}, then there would be a column called \code{party_id} containing the
values of that variable in long form.
\item \code{pr_pred}: The predicted probability of taking the value of the outcome. This
is the main output of the multinomial model.
\item \code{count}: A new count variable. Simply the product of \code{n_aggregate} and \code{pr_pred}.
}
}
\description{
Imputes the counts of a joint distribution of count variables for small areas
based on microdata. See Details.
}
\details{
In this setup, the population distribution table (\code{poptable})
has the joint distribution of \verb{(A, X_\{1\}, ..., X_\{K - 1\})} categorical variables where
\code{A} denotes a categorical small area, \code{X}s denote categorical covariates, and
the missing covariate is \verb{X_\{K\}}.

Now, the survey data (\code{microdata}) has a sample joint distribution of
\verb{(X_\{1\}, .., X_\{K - 1\}, X_\{K\})} categorical variables but the sample size is too small
for small areas. Therefore, the function models a multinomial outcome model roughly of the
form \verb{X_\{K\} ~ X_\{1\} + ... X_\{K\}} and predicts onto \code{poptable} to estimate the
joint distribution of \verb{(A, X_\{1\}, .., X_\{K\})}

Currently, this function does not support post-stratifiation based on a known aggregate
distribution -- that is, further adjusting the probabilities based on
a known population distribution (see e.g. Leeman and Wasserfallen AJPS \url{https://doi.org/10.1111/ajps.12319})
}
\examples{
 library(dplyr)

 # Impute the joint distribution of party ID with race, sex, and age, using
 # survey data in NY.

 imputed_acs <- synth_mlogit(pid3 ~ race + age + female,
                             microdata = cc18_NY,
                             poptable = acs_NY,
                             area_var = "cd")

 # original (27 districts x 2 sex x 5 age x 6 race categories)
 count(acs_NY, cd, female, age, race, wt = count)

 # new, modeled (original x 5 party categories)
 imputed_acs

 # See the data elec_NY to see if these numbers look reasonable.


\dontrun{
  # another example -- imputing education
  library(ccesMRPrun)
  synth_mlogit(educ ~ age + female,
              microdata = cces_GA,
              poptable = acs_GA,
              area_var = "cd")
}

}
